[optimizer] don't merge "A+B|B => A?+B" for `ChoiceRule`
we actually can't do this.

for `A+B1|B2`:
   - when `A` matches, tries to match `B1`.
      - when `B1` matches, we found a match for `A+B1` and advance.
      - when `B1` fails, `A+B1` fails, fallback to `B2`
   - when `A` fails, try to match B2.

however, for `A?+B`:
   - when `A` matches, tries to match `B`
   - when `A` fails, tries to match `B`

the problem is for `A?+B`, when `A` matches, it wants to match a `B`, and success only when a valid `B` matches.
if `A` matches but `B` doesn't, `A?+B` fails.
this actually behaves different than `A+B|B`.